# app/swagger-2.0.yaml
swagger: '2.0'
info:
  title: Aichat API
  version: '1.0.0'
  description: |
    当前项目新接口的 Swagger 2.0 文档。

    说明：
    - 所有接口统一走 `/api/...`
    - 除 `/api/login` 外，其余接口默认需要 token
    - 推荐使用 `Authorization: Bearer <token>`
host: localhost
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: '格式：Bearer <token>'
paths:
...
  /api/download/file:
    post:
      tags:
        - Download
      summary: 下载指定文件
      description: 按文件名或文件 ID 下载文件，成功后更新下载状态。
      security:
        - BearerAuth: []
      produces:
        - application/octet-stream
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              filename:
                type: string
                description: 文件名，与 file_id 至少传一个
              file_id:
                type: integer
                description: 文件 ID，与 filename 至少传一个
      responses:
        '200':
          description: 下载成功，返回文件流
          schema:
            type: string
            format: binary
...
  /api/download/iec/file:
    post:
      tags:
        - Download
      summary: 下载 IEC 文件
      description: 返回固定 IEC 文件流。
      security:
        - BearerAuth: []
      produces:
        - application/octet-stream
      responses:
        '200':
          description: 下载成功
          schema:
            type: string
            format: binary
...